instruct-insertion/model/referit3d/external_tools/Scan2CAD/Routines/Script/requirements.txt
numpy
numpy-quaternion
plyfile
pybind11
PyWavefront
